requirements.txt
flask
flask-cors
mysql-connector-python
rpy2
pandas
numpy
python-dotenv